relocate tests/ benchmarks to different folder - change directory names
